[Identity] Update example app
