--- identity-example/src/theme2/java/com/stripe/android/identity/example/ThemeActivity.kt
@@ -1,5 +1,0 @@
-package com.stripe.android.identity.example
-
-internal class ThemeActivity : MainActivity() {
-    override val getBrandLogoResId = R.drawable.merchant_logo_red
-}
